그리드 헤더, 컬럼 헤더 구분 스토리 추가/수정.
