+++ stories/04-GridHeader/onlydel.gridview.header.stories.js
@@ -5,7 +5,7 @@ import init from '../_common/init';
 import { renderMessage } from '../_common/utils';
 
 export default {
-  title: '04-헤더',
+  title: '04-그리드 헤더',
 };
 
 export const 헤더감추기 = () => {
@@ -27,7 +27,7 @@ export const 헤더높이_60 = () => {
 
     setTimeout(() => {
       const header = gridView.getHeader();
-      renderMessage(header.height === 61 ? '헤더높이 60 맞네' : new Error('헤더높이 60 아닌데'));
+      renderMessage(header.height === 60 ? '헤더높이 60 맞네' : new Error('헤더높이 60 아닌데'));
     }, 100);
   });
 
@@ -42,19 +42,6 @@ export const 헤더속성_기본값 = () => {
       const header = gridView.getHeader();
       renderMessage(JSON.stringify(header));
     }, 100);
-  });
-
-  return container;
-};
-
-export const 헤더체크컬럼 = () => {
-  useEffect(() => {
-    const { gridView } = init('realgrid');
-
-    const checkColumn = gridView.columnByName('KorName');
-    checkColumn.header.checkLocation = 'left';
-    checkColumn.checked = true;
-    gridView.setColumn(checkColumn);
   });
 
   return container;
